@@ -79,32 +94,36 @@
     "toml": "^2.2.2",
     "toml-loader": "^1.0.0",
     "tracer": "^0.8.3",
-    "typography": "^0.13.0",
+    "typography": "^0.13.1",
+    "unist-util-select": "^1.5.0",
     "url-loader": "^0.5.7",
     "webpack": "^1.13.2",
     "webpack-configurator": "^0.3.0",
     "webpack-hot-middleware": "^2.12.2",
+    "webpack-md5-hash": "0.0.5",
     "webpack-require": "0.0.16",
+    "webpack-stats-plugin": "^0.1.3",
+    "yaml-js": "^0.1.3",
     "yaml-loader": "^0.4.0"
   },
   "devDependencies": {
     "ava": "^0.16.0",
     "ava-http": "^0.2.1",
-    "babel-cli": "^6.11.4",
+    "babel-cli": "^6.14.0",
     "babel-eslint": "^6.1.2",
     "babel-plugin-lodash": "^3.2.8",
-    "babel-plugin-transform-flow-strip-types": "^6.8.0",
-    "babel-register": "^6.11.6",
-    "bluebird": "^3.4.1",
+    "babel-plugin-transform-flow-strip-types": "^6.14.0",
+    "babel-register": "^6.14.0",
+    "bluebird": "^3.4.2",
     "cheerio": "^0.20.0",
-    "eslint": "^2.11.1",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-ava": "^2.5.0",
+    "eslint": "^3.3.1",
+    "eslint-config-airbnb": "^10.0.1",
+    "eslint-plugin-ava": "^3.0.0",
     "eslint-plugin-flow-vars": "^0.5.0",
-    "eslint-plugin-flowtype": "^2.7.1",
-    "eslint-plugin-import": "^1.13.0",
-    "eslint-plugin-jsx-a11y": "1.5.3",
-    "eslint-plugin-react": "^5.2.2",
+    "eslint-plugin-flowtype": "^2.11.1",
+    "eslint-plugin-import": "^1.14.0",
+    "eslint-plugin-jsx-a11y": "^2.1.0",
+    "eslint-plugin-react": "^6.1.2",
     "flow-bin": "^0.30.0",
     "iflow-debug": "^1.0.15",
     "iflow-lodash": "^1.1.23",
