@@ -2,35 +2,37 @@
 import React from 'react'
 import { renderToString, renderToStaticMarkup } from 'react-dom/server'
 import { match, RouterContext } from 'react-router'
-import createRoutes from 'create-routes'
+//import createRoutes from 'create-routes'
 import Html from 'html'
-const loadContext = require('.gatsby-context')
-import { pages, config } from 'config'
-import { prefixLink } from '../isomorphic/gatsby-helpers'
-
-let routes
-loadContext((pagesReq) => {
-  routes = createRoutes(pages, pagesReq)
-})
+import _ from 'lodash'
+//import { pages, config } from 'config'
+//import { prefixLink } from '../isomorphic/gatsby-helpers'
+//import { routesDB } from './globals'
+import childRoutes from '.gatsby-intermediate-build/child-routes.js'
+const routes = {
+  childRoutes,
+}
 
 module.exports = (locals, callback) => {
-  match({ routes, location: prefixLink(locals.path) }, (error, redirectLocation, renderProps) => {
+  match({ routes, location: locals.path }, (error, redirectLocation, renderProps) => {
     if (error) {
-      console.log(error)
+      console.log(`error when building page ${locals.path}`, error)
       callback(error)
     } else if (renderProps) {
-      const component = <RouterContext {...renderProps} />
-      let body
-
-      // If we're not generating a SPA for production, eliminate React IDs.
-      if (config.noProductionJavascript) {
-        body = renderToStaticMarkup(component)
-      } else {
-        body = renderToString(component)
-      }
+      const component = (
+        <RouterContext
+          {...renderProps}
+        />
+      )
+      const body = renderToString(component)
 
       const html = `<!DOCTYPE html>\n ${renderToStaticMarkup(
-        <Html body={body} {...renderProps} />
+        <Html
+          body={body}
+          {...renderProps}
+          pathJSFile={`path---${locals.path === '/' ? 'index' : _.kebabCase(locals.path)}`}
+          componentName={_.find(childRoutes, (child) => child.path === locals.path).componentName}
+        />
       )}`
       callback(null, html)
     }
