@@ -4,11 +4,11 @@ const fs = require('fs-extra')
 const path = require('path')
 import _ from 'lodash'
 
-// Copy our load-context function to root of site in a dot file.
-const gatsbyFile = `${__dirname}/../utils/load-context.js`
-const siteDirectory = path.resolve('.')
-const fileName = `${siteDirectory}/.gatsby-context.js`
-fs.copy(gatsbyFile, fileName)
+// Copy our site files to the root of the site.
+const srcDir = `${__dirname}/../intermediate-build-dir`
+const siteDirectoryRoot = path.resolve('.')
+const siteDir = `${siteDirectoryRoot}/.gatsby-intermediate-build`
+//fs.copy(srcDir, siteDir, (err) => { if (err) { console.log(err) }})
 
 const defaultHost = process.platform === 'win32'
   ? 'localhost'
