@@ -1,5 +1,5 @@
 /* @flow weak */
-import { config } from 'config'
+import { linkPrefix } from 'gatsby-browser'
 import invariant from 'invariant'
 import _ from 'lodash'
 
@@ -14,20 +14,20 @@ function isDataURL (s) {
 const prefixLink = (_link) => {
   if (
       (typeof __PREFIX_LINKS__ !== 'undefined' && __PREFIX_LINKS__ !== null)
-      && __PREFIX_LINKS__ && (config.linkPrefix !== null
+      && __PREFIX_LINKS__ && (linkPrefix !== null
   )) {
     const invariantMessage = `
     You're trying to build your site with links prefixed
     but you haven't set 'linkPrefix' in your config.toml.
     `
-    invariant(_.isString(config.linkPrefix), invariantMessage)
+    invariant(_.isString(linkPrefix), invariantMessage)
 
     // When url-loader is used with the limit option, some of images are encoded to
     // base64 data url, this check prevent such links from being prefixed.
     if (isDataURL(_link)) {
       return _link
     } else {
-      return config.linkPrefix + _link
+      return linkPrefix + _link
     }
   } else {
     return _link
