@@ -7,7 +7,9 @@ import buildCSS from './build-css'
 import buildHTML from './build-html'
 import buildProductionBundle from './build-javascript'
 import postBuild from './post-build'
-import globPages from './glob-pages'
+//import globPages from './glob-pages'
+import bootstrap from './bootstrap'
+import { routesDB } from './globals'
 
 function customPost (program, callback) {
   const directory = program.directory
@@ -25,15 +27,13 @@ function customPost (program, callback) {
   if (customPostBuild) {
     console.log('Performing custom post-build steps')
 
-    return globPages(directory, (globError, pages) =>
-      customPostBuild(pages, (error) => {
-        if (error) {
-          console.log('customPostBuild function failed')
-          callback(error)
-        }
-        return callback()
-      })
-    )
+    customPostBuild(routesDB.data, (error) => {
+      if (error) {
+        console.log('customPostBuild function failed')
+        callback(error)
+      }
+      return callback()
+    })
   }
 
   return callback()
@@ -60,8 +60,7 @@ function bundle (program, callback) {
       console.log('failed to compile bundle.js')
       return callback(error)
     }
-
-    return post(program, callback)
+    return callback()
   })
 }
 
@@ -75,26 +74,28 @@ function html (program, callback) {
     callback(error)
   }
 
-  console.log('Generating CSS')
-  buildCSS(program, (cssError) => {
-    if (cssError) {
-      console.log('Failed at generating styles.css')
-      return callback(cssError)
-    }
-
-    console.log('Generating Static HTML')
-    return buildHTML(program, (htmlError) => {
-      if (htmlError) {
-        console.log('Failed at generating HTML')
-        return callback(htmlError)
+  return bootstrap(program, (err, routes) => {
+    console.log('Generating CSS')
+    buildCSS(program, (cssError) => {
+      if (cssError) {
+        console.log('Failed at generating styles.css')
+        return callback(cssError)
       }
 
-      // If we're not generating javascript, go directly to post build steps
-      if (config.noProductionJavascript) {
-        return post(program, callback)
-      } else {
-        return bundle(program, callback)
-      }
+      return bundle(program, (javascriptError) => {
+        if (javascriptError) {
+          console.log('Failed at building Javascript bundles')
+          return callback(javascriptError)
+        }
+        console.log('Generating Static HTML')
+        return buildHTML(program, (htmlError) => {
+          if (htmlError) {
+            console.log('Failed at generating HTML')
+            return callback(htmlError)
+          }
+          return post(program, callback)
+        })
+      })
     })
   })
 }
