@@ -15,13 +15,15 @@ import opn from 'opn'
 import fs from 'fs'
 import glob from 'glob'
 import rl from 'readline'
+import GraphQL from 'hapi-graphql'
+
+import bootstrap from './bootstrap'
 
 const rlInterface = rl.createInterface({
   input: process.stdin,
   output: process.stdout,
 })
 
-import globPages from './glob-pages'
 import webpackConfig from './webpack.config'
 const debug = require('debug')('gatsby:application')
 
@@ -30,7 +32,8 @@ function startServer (program, launchPort) {
   const serverPort = launchPort || program.port
 
   // Load pages for the site.
-  return globPages(directory, (err, pages) => {
+  return bootstrap(program, (err, routes, schema) => {
+  //return globPages(directory, (err, pages) => {
     const compilerConfig = webpackConfig(program, directory, 'develop', program.port)
 
     const compiler = webpack(compilerConfig.resolve())
@@ -98,17 +101,21 @@ function startServer (program, launchPort) {
       })
 
       server.ext('onRequest', (request, reply) => {
+        if (request.path === '/graphql') {
+          return reply.continue()
+        }
+
         const negotiator = new Negotiator(request.raw.req)
 
         // Try to map the url path to match an actual path of a file on disk.
         const parsed = parsePath(request.path)
-        const page = _.find(pages, (p) => p.path === (`${parsed.dirname}/`))
+        const page = _.find(routes, (p) => p.path === (`${parsed.dirname}/`))
 
         let absolutePath = `${program.directory}/pages`
         let path
         if (page) {
-          path = `/${parsePath(page.requirePath).dirname}/${parsed.basename}`
-          absolutePath += `/${parsePath(page.requirePath).dirname}/${parsed.basename}`
+          path = `/${parsePath(page.component).dirname}/${parsed.basename}`
+          absolutePath += `/${parsePath(page.component).dirname}/${parsed.basename}`
         } else {
           path = request.path
           absolutePath += request.path
@@ -153,14 +160,31 @@ function startServer (program, launchPort) {
         },
       }
 
-      return server.register({
-        register: WebpackPlugin,
-        options: {
-          compiler,
-          assets,
-          hot,
+
+      return server.register([
+        // Add GraphQL support
+        {
+          register: GraphQL,
+          options: {
+            query: {
+              graphiql: true,
+              pretty: true,
+              schema,
+            },
+            route: {
+              path: '/graphql',
+            },
+          }
         },
-      }, (er) => {
+        {
+          register: WebpackPlugin,
+          options: {
+            compiler,
+            assets,
+            hot,
+          },
+        },
+      ], (er) => {
         if (er) {
           console.log(er)
           process.exit()
@@ -189,6 +213,10 @@ function startServer (program, launchPort) {
 }
 
 module.exports = (program) => {
+  // TODO boot checker — gatsby.config.js/html.js/local install
+
+  // Detect if a process is already listening on the program port.
+  // If it is, ask the user if they'd like to use a different port.
   detect(program.port, (err, _port) => {
     if (err) {
       console.error(err)
