@@ -6,6 +6,10 @@ const debug = require('debug')('gatsby:webpack-config')
 import path from 'path'
 import _ from 'lodash'
 import invariant from 'invariant'
+import { StatsWriterPlugin } from "webpack-stats-plugin"
+const WebpackMD5Hash = require('webpack-md5-hash')
+const OfflinePlugin = require('offline-plugin')
+const ChunkManifestPlugin = require('chunk-manifest-webpack-plugin')
 
 import babelConfig from './babel-config'
 
@@ -25,7 +29,7 @@ try {
 //   2) develop-html: same as develop without react-hmre in the babel config for html renderer
 //   3) build-css: build styles.css file
 //   4) build-html: build all HTML files
-//   5) build-javascript: Build bundle.js for Single Page App in production
+//   5) build-javascript: Build js chunks for Single Page App in production
 
 module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes = []) => {
   const babelStage = suppliedStage
@@ -37,7 +41,7 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
       case 'develop':
         return {
           path: directory,
-          filename: 'bundle.js',
+          filename: '[name].js',
           publicPath: `http://${program.host}:${webpackPort}/`,
         }
       case 'build-css':
@@ -58,8 +62,10 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
         }
       case 'build-javascript':
         return {
-          filename: 'bundle.js',
+          //filename: '[name].js',
+          filename: '[name]-[chunkhash:8].js',
           path: `${directory}/public`,
+          publicPath: '/frontend/',
         }
       default:
         throw new Error(`The state requested ${stage} doesn't exist.`)
@@ -69,22 +75,33 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
   function entry () {
     switch (stage) {
       case 'develop':
-        return [
-          require.resolve('webpack-hot-middleware/client'),
-          `${__dirname}/web-entry`,
-        ]
+        return {
+          commons: [
+            require.resolve('webpack-hot-middleware/client'),
+            `${directory}/.gatsby-intermediate-build/routes`,
+          ],
+        }
       case 'build-css':
         return {
-          main: `${__dirname}/web-entry`,
+          main: `${directory}/.gatsby-intermediate-build/routes`,
         }
       case 'build-html':
         return {
           main: `${__dirname}/static-entry`,
         }
       case 'build-javascript':
-        return [
-          `${__dirname}/web-entry`,
-        ]
+        return {
+          //vendor: ["react", "react-dom", "react-router", "typography", 'react-responsive-grid', 'react-document-title'],
+          //index: [`${directory}/pages/index`],
+          //blogPost: [`${directory}/pages/blog-post`],
+          //tagPage: [`${directory}/pages/tag-page`],
+          //tagsPage: [`${directory}/pages/tags-page`],
+          routes: `${directory}/.gatsby-intermediate-build/split-routes`,
+          // TODO create new chunks per *route* to soak up libs specific to each route
+          // so each page chunk restricted to just data needed for that page.
+          // Then "vendor" tweaking won't be needed as much.
+          // TODO inline JS files < 10 kb?
+        }
       default:
         throw new Error(`The state requested ${stage} doesn't exist.`)
     }
@@ -131,6 +148,12 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
           // by default in Webpack that most sites don't want.
           // This line disables that.
           new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
+          new WebpackMD5Hash(),
+          new webpack.optimize.DedupePlugin(),
+          new webpack.optimize.CommonsChunkPlugin({
+            name: 'commons',
+            minChunks: 2,
+          }),
           new webpack.DefinePlugin({
             'process.env': {
               NODE_ENV: JSON.stringify(process.env.NODE_ENV ? process.env.NODE_ENV : 'production'),
@@ -138,8 +161,22 @@ module.exports = (program, directory, suppliedStage, webpackPort = 1500, routes 
             __PREFIX_LINKS__: program.prefixLinks,
           }),
           new ExtractTextPlugin('styles.css'),
-          new webpack.optimize.DedupePlugin(),
+          //new OfflinePlugin({
+            ////AppCache: false,
+            //publicPath: '/',
+            //relativePaths: false,
+            //ServiceWorker: {
+              //events: true
+            //},
+          //}),
           new webpack.optimize.UglifyJsPlugin(),
+          new StatsWriterPlugin(),
+          //new ChunkManifestPlugin({
+            //filename: "chunk-manifest.json",
+            //manifestVariable: "webpackManifest"
+          //}),
+          new webpack.optimize.OccurenceOrderPlugin(),
+          // inline manifest in head, atom.xml as "page"
         ]
       default:
         throw new Error(`The state requested ${stage} doesn't exist.`)
